♻️ refactor(client): move stableAvatarColorHex to core; guard data→presentation imports
